+++ laokou-common/laokou-common-data-cache/src/main/java/org/laokou/common/data/cache/aspectj/package-info.java
@@ -18,4 +18,4 @@
 /**
  * @author laokou
  */
-package org.laokou.common.secret.aop;
+package org.laokou.common.data.cache.aspectj;
